feat: Curate, scrape and add articles to database
This brings everything together; it reads from the registered Sources in
the db, curates the articles from there, scrapes the content for NLP
tasks and creates Article items.

Although not implemented yet, the idea is to have this run every 24h.
Meaning that our database would always have the latest news.

Additionally, by using the Sources table this way, it allows for users
to add sources they want the cron job to use.