feat: Ensure rating in query is always latest
- Add useEffect for better state update handling
- Address issue where updateDbRating was always using the previous
rating instead of the latest

Before, due to how React queues re-rendering and state updates, the
`updateDbRating` function was always using the previous rating, rather
than the current one.

Now, with the use of useEffect, `updateDbRating` will use the latest rating.

An alternative considered was to place `updateDbRating` inside the
`setState` function. Although this would have solved the issue, it
wouldn't align well with the React philosophy of avoiding side effects
in state update functions.